refactor(misc1): move preserve_exit() to related functions in main.c
