+++ src/nvim/event/rstream.c
@@ -12,6 +12,7 @@
 #include "nvim/event/rstream.h"
 #include "nvim/log.h"
 #include "nvim/memory.h"
+#include "nvim/main.h"
 #include "nvim/misc1.h"
 #include "nvim/vim.h"
